Added an aerial viewpoint and target viewpoints for all of the buildings in the scene. When we first show the town view, we sweep down to the main viewpoint from the aerial viewpoint, then we go from there to a particular building's viewpoint when we're about to enter.
The broad strokes are in place, but I need to refine the timing, and I'm not
sure a SplinePath is the ideal way to manipulate the camera as it yanks the
view around pretty rapidly at times. I also need to display the town name when
we're doing our "presentation" of the town.