Step one: Start implementing top 10 lists in the saloon.
Step two: Decide it would be really cool if we showed the avatar of the number
one player in each list.

Step three: Realize that means we need to store "snapshots" of players' avatars
because we otherwise can't reconstruct them without loading up their inventory
which would suck.

Step four: Decide that it's time to implement the whole "assign a specific look
to each of a set of 'poses' functionality" so that we can have a "Wanted
Poster" pose which will be snapshotted and made available for viewing when the
player is offline.

Still need to actually save the snapshots and load them at the appropriate
time, and create an interface for assigning looks to poses, but all the basic
bits are now in place and the top 10 lists work other than showing the avatar
of the number one player.