Rate games that have gone on for three minutes instead of four. I think it's a better balance of the following two situations:
- On the one hand, assholes could try to game their ratings by playing against
a mule account for as little time as possible, then having that account
disconnect, ending the game in their favor. So we want to force them to play
for as long as we can to make it difficult to do that.

- On the other hand, assholes can disconnect from a real game as soon as they
think they are going to lose, thereby screwing the other player who was
legitimately trying to play. So we want to prevent them from doing that by
rating the game as soon as they have played any meaningful amount of it.

Oh the joy of balancing one asshole move against another.