Made some changes to initiative to approximate what will likely happen when that is assigned dynamically.
Made a change to the Old Codger so that he does not do 100 damage against all
air units. He does 85 to all air units and 100 to steam air units. This ups his
damage versus the Artillery and Steam Gunman from 50 to 65 which I think is
reasonable and it makes him less of a powerhouse against the two new air units
in Indian Trading Post.

Added first pass unit properties for the Storm Caller and Trickster Raven.  All
the ITP units will need their bonuses and penalties tweaked but we'll do that
later.

Added a header to each properties file so one can easily tell what in the hell
they're looking at.