The claim -> counter refactor broke nugget handling in the tutorial (which didn't extend GoldScenario and thus didn't have proper NuggetEffect handling). This prompted me to do the giant Scenario refactor that I'd been wanting to do wherein each individual scenario behavior was factored out into a delegate and then the main scenario could mix and match those to create the behavior it needed.
We could almost do away with the Scenario entirely and have each scenario
simply be a collection of ScenarioDelegates but there are a few things that are
nice to have each Scenario derivation, so we'll go with this.

I also added a steam tank to the gold rush tutorial and changed its text to
explain the steam tank. And I added rustling posts to the cattle rustling
scenario, but I did not add extra text to explain them.

The fact that the rustling post counters only update on each tick results in an
annoying delay after a cattle is branded before they update. It's also kind of
silly to broadcast this information from the server when the client could just
as easily calculate it from known data, but we'll save that fixing for later.