Effects are now reported as:
  town_id/effect_identifier

and we use that base effect name to look up sound effects in a variety of
places (bonuses, cards, extras, effects) and we look for an effect with the
specified name. Thus sometimes we just use the name of a bonus:

  frontier_town/hustle

other times the name of a card:

  frontier_town/giddy_up

other times the name of a particle system effect:

  indian_post/storm_caller/attack

The sound searching does fancy stuff, the effect searching just slaps effects/
on the front and checks to see if we have an effect with that name.

Some stuff might be broken, so try to keep an eye/ear out.