New bonus placement algorithm, we pick a unit at random, weighted by the scores of the players (more likely to pick someone in last place, zero chance of picking first place player's units unless everyone is tied).
Then pick a spot in that unit's movement range and place the bonus there using
the existing algorithm to determine what sort of bonus to place based on who
can reach that spot (though that algorithm could use some tweaking since it
only considers units that are ready to move immediately which are rare).

Early testing felt like there were a damned lot of bonuses in play which was
good, but I toned it down so that it only tries to keep one fewer than the
total number of players in the game bonuses on the board.