Refuse to include any server code in the Proguarded client. Nothing should reference anything from the server packages. The editor references some things from presents/server and parlor/server but nothing from Bang's server code and should not ever need to do so (meaning we should put anything needed by the client or editor and the server into a shared util class).
It would be cool if we could somehow "ignore" the server .class files when
running the editor or client from the command line because then we wouldn't
have to find out something is broken because of such a reference on the dev
build. I'm not sure if such a thing is possible without just copying everything
but the server code into a separate directory and using that instead of
dist/classes.