Bah. We now have to immediately switch watchers to the next board because there is no coordinated distributed event that takes place between the end of the previous round and the start of the next (by which time it is too late to switch to the new board).
Players automatically switch when they dismiss the stats and move to selecting
their team and the next round won't start until they've all done so. However,
watchers could blissfully leave their stats window up well into the next round
so we can't depend on that as a trigger for the board transition.