More work on bounties. Everything is in place now to design and test bounty games. Criterion can be specified and they are checked at the end of the game (and reported via chat for the time being).
I also revamped the way Forest Guardians tracks perfect waves so that that stat
is available in the game object statistics sets. In the proces I nixed the
assignment of a perfect wave for the incomplete final wave because it was
possible to simply not grow any trees at all and that partial wave would be
declared perfect.