More revamping, more progress.
BangConfig now contains Round records that specify the configuration of each
round (the scenario, a specific board if desired and specific board data if
desired). This allowed all the board selection and decoding logic to move into
BoardManager, cleaning up BangManager nicely, and it will allow us, if we
desire, to have multiple round bounty games.

Made BangConfig and the associated bits Savable so that we can save and load
bounty game configurations in a future compatible way.

Next up is adding loading and saving to the editor (for testing) and coming up
with a way for us to easily share bounty configurations during testing.