I know this seems strange, but we don't want to update Bang to track changes in Narya, etc. It uses versioned dependencies now and is happily living in its little snasphot of history.
This allows us to fix bugs in Bang without taking on the burden of ensuring
that all of the upstream changes didn't happen to break anything.

As it happens, Bang is sitting on a year or more of upstream changes made
before it was possible to convert it to live in the past, so anyone intrepid
enough to ship a Bang release will have to deal with that heap of potential
breakage. However, once those niggles are sorted out, Bang could potentially
ship bug fix releases in full confidence that no surprises were lurking in the
shadows, waiting to wreak havoc.