Oh the huge manatee. Now we're back to ant fucking us because distcleanall depends on tasks that initialize ant-contrib, and then when we call bamboo, it tries to reinitialize ant-contrib and chokes. Fucking ant.
Our new plan is that we *will* assume a clean build, and we'll distcleanall and
checkclean at the very end of our build to make sure that we do leave things in
a pristine state at the end of every (successful) build. If a build fails,
we're just going to leave the crap lying around and the next build can
overwrite it. Given that this is how we do things all the time when doing
incremental builds on developer machines, I'm not worried that this is going to
enable some terrible inter-build dependency to sneak past the guards.

On the plus side, this will also avoid all the annoying problems where removing
an svn:external causes future builds to break because they're "not clean",
because the build didn't clean up after itself and the external failed to
completely disappear because there was some cruft inside it from a previous
build.